Constrain attached window positions
Limit the positions to which the windows can be attached to those that
leave at least a small margin (currently 40px width, 40px height incl.
decorations) of the window visible. This is done to avoid windows
attaching extremely far if they're on a different workspace / monitor,
as well as hopefully help the user avoid losing their windows in
general.