Use combine() for shortcuts instead of | or +
While the documentation suggests using + (or | for Qt 6) for combining
key sequences, keys and modifier keys are of different enum types, which
raises the deprecated-enum-enum-conversion warning, which results in a
failed compilation due to our use of -Werror. Fun!

This isn't the first time this exact thing has been a problem, which is
why combine() in include/DeprecationHelper.h exists. This commit uses
that.

https://doc.qt.io/archives/qt-5.15/qkeysequence.html#details
https://doc.qt.io/qt-6/qkeysequence.html#details