Fix effect control views bing killed on close
DeleteOnClose flag is unset from being default in SubWindow since it
doesn't seem to be a desired default and in many places simply requires
workarounds that disable it.

Useless size restrictions EffectView imposes on SubWindow are removed
since those are already implicit and conform to layout size policy of
controlView.