Remove `fastMemCpy()` from RemotePluginBase
All `fastMemCpy()` does is check if the size is equal to 4 bytes, in
which case it casts the pointers to `int32_t*` and does an assignment
instead of calling into `memcpy()`. Otherwise, it calls `memcpy()`
normally.

This is an anti-optimization as it gets in the way of all the normal
`memcpy()` optimizations that compilers normally do.