Add adaptive EMA scaling
Early taps (as well as taps with high interval variance in general) need more responsive BPM updates so we can get to a target range quickly. Once we have reached the target BPM range, we want the BPM readings to become more stable as the user trying to tap with more consistency.

We use EMA to reduce jitter in the BPM readings, but the amount of smoothing we apply should change dynamically to align with how the user is tapping.

Adaptive EMA scaling adjusts the smoothing factor dynamically to account for this based on how much the tap intervals vary. If the variation is high, then the BPM updates are more responsive (allowing for rapid changes). If the variation is low, then the BPM updates are less responsive (slower, more stable changes).